refactor:重构部分代码

+++ mooc-message/src/main/java/cn/edu/lingnan/mooc/message/controller/NoticeController.java
@@ -95,7 +95,7 @@ public class NoticeController {
      * @return
      */
     @GetMapping("/reply/detail")
-    public RespResult getReplyMessageDetail(@RequestParam Integer id){
+    public RespResult getReplyMessageDetail(@RequestParam Long id){
         return RespResult.success(noticeService.getReplyNoticeDetail(id));
     }
